OK; three working, fairly quick, integration tests.
These need to become actual tests. i.e. include a check that the energy
is vaguely sensible, in the limits of the MC runs we can sensibly put
within the testing environment.

It might be nice to have an earlier test that literally just iterates
through the possible simulation configurations, for a bare minimum of MC
moves so that you can see the code failing if you mess up the types etc.

Perhaps harmonic -> Frohlich -> Holstein will suffice as escalatingly
complex, particularly once we add multiple particles.